+++ src/main/java/frc/robot/commands/Intake.java
@@ -10,7 +10,7 @@ import frc.robot.subsystems.LaunchSubsystem;
 /** An example command that uses an example subsystem. */
 public class Intake extends CommandBase {
   private final LaunchSubsystem launchSubsystem;
-  private float intakeSpeed = -0.25f;
+  private double intakeSpeed = -0.25;
 
   public Intake(LaunchSubsystem subsystem) {
     launchSubsystem = subsystem;
